pyproject.toml
[project]
name = "csc316-final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Luna", email = "itsactuallyluna9@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "numpy>=2.3.3",
    "obsws-python>=1.8.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.0.0",
    "pynput>=1.8.1",
    "rich>=14.1.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "ultralytics>=8.3.200",
]

[project.scripts]
csc316-final-project = "csc316_final_project:cli"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"
